+++ comics/comics/dungeond.py
@@ -12,6 +12,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 365
+    schedule = 'Tu'
     time_zone = 'US/Pacific'
 
     def crawl(self, pub_date):