+++ comics/comics/hijinksensue.py
@@ -11,11 +11,11 @@ class ComicData(ComicDataBase):
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 28
+    history_capable_days = 40
     time_zone = 'US/Central'
 
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://feeds.feedburner.com/hijinksensue')
+        feed = self.parse_feed('http://hijinksensue.com/feed/')
         for entry in feed.for_date(pub_date):
             if 'Comics' not in entry.tags:
                 continue