+++ comics/comics/johnnywander.py
@@ -12,7 +12,6 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_days = 40
-    schedule = 'Tu,Th'
     time_zone = 'US/Pacific'
 
     def crawl(self, pub_date):