+++ comics/comics/pelsogpoter.py
@@ -12,7 +12,7 @@ class ComicData(ComicDataBase):
 
 class Crawler(CrawlerBase):
     history_capable_date = '2009-03-23'
-    schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
+    schedule = 'Mo,We,Th,Fr,Sa,Su'
     time_zone = 'Europe/Oslo'
 
     def crawl(self, pub_date):