+++ comics/comics/sequentialart.py
@@ -11,6 +11,7 @@ class ComicData(ComicDataBase):
 
 
 class Crawler(CrawlerBase):
+    schedule = 'We'
     time_zone = 'Europe/London'
 
     def crawl(self, pub_date):